% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsis_control.R
\name{apsis_control}
\alias{apsis_control}
\title{Auxiliary Function for Controlling Parameters in APSIS Algorithm}
\usage{
apsis_control(epsilon = 1e-04, max.iter = 1000, step.size = NULL)
}
\arguments{
\item{epsilon}{A numeric value. the tolerance being used to check the
absolute differences between two generated parameters from target
distribution and stop the simulation. The default value is 1e-3.}

\item{max.iter}{A positive numeric value. The maximum number of Monte Carlo
simulation that you want the algorithm to take. The default value is 1000.
The algorithm stops when either desired tolerance achieved or max number of
iteration reached, whichever comes first.}

\item{step.size}{A positive numeric value for step size value when updating
proposal. The best practical suggestion is to set stepsize as \frac{1}{t}
where t is the iteration number and this is the default behaviour when
step.size = NULL.}
}
\value{
A list of three.
}
\description{
Auxiliary Function for Controlling Parameters in APSIS Algorithm
}
