% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k_hat.R
\name{get_k_hat}
\alias{get_k_hat}
\title{Estimate k for a given value of theta and a given dataset}
\usage{
get_k_hat(theta, x)
}
\arguments{
\item{X}{}
}
\description{
Estimate k for a given value of theta and a given dataset
}
