% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thetas.R
\name{get_thetas}
\alias{get_thetas}
\title{Compute intermediate estimates of theta for fitting the Pareto distribution}
\usage{
get_thetas(X_star, X_max, M_theta)
}
\arguments{
\item{M_theta}{}
}
\description{
Compute intermediate estimates of theta for fitting the Pareto distribution
}
