% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_smoothing.R
\name{pareto_smoothing}
\alias{pareto_smoothing}
\title{Apply Pareto Smoothing to Weights}
\usage{
pareto_smoothing(X, M_keep, return_k = FALSE, regularize = TRUE)
}
\arguments{
\item{regularize}{}
}
\description{
Apply Pareto Smoothing to Weights
}
