% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_proposal_normal.R
\name{optimise_proposal_normal}
\alias{optimise_proposal_normal}
\title{Pareto Smoothed Importance Sampling with a Family of Normal Distribution For
Proposal}
\usage{
optimise_proposal_normal(target, n, mu = c(-10, 10), control = NULL)
}
\arguments{
\item{mu}{}
}
\description{
Pareto Smoothed Importance Sampling with a Family of Normal Distribution For
Proposal
}
